Move class to a standalone file.
